# packages/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AI-HRMS API
  version: 0.1.0
  description: >
    Contract index for the current AI-HRMS reproduction phase. The API serves
    one corporate group with multiple legal entities; legal entities are
    organization scopes, not SaaS tenants.
servers:
  - url: http://localhost:8080/api
    description: Local Go API
security:
  - bearerAuth: []
...
paths:
...
  /rag/ingest-jobs/{id}:
    get:
      tags: [RAG]
      operationId: getRagIngestJob
      summary: Get ingestion job
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Ingestion job.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
  /learning/courses/{id}/lessons:
    get:
      tags: [Learning]
      operationId: listLearningLessons
      summary: List lessons for a course
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Lesson list.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiEnvelope"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ApiEnvelope:
      type: object
      required: [success, code, message]
      properties:
        success:
          type: boolean
        code:
          type: integer
          example: 2001
        message:
          type: string
        data: {}
...